Fully renamed app to Web3Modal
