+++ dapps/v2Explorer/android/app/build.gradle
@@ -93,9 +93,9 @@ android {
 
     compileSdkVersion rootProject.ext.compileSdkVersion
 
-    namespace "com.v2explorer"
+    namespace "com.walletconnect.web3modal.rnsample"
     defaultConfig {
-        applicationId "com.v2explorer"
+        applicationId "com.walletconnect.web3modal.rnsample"
         minSdkVersion rootProject.ext.minSdkVersion
         targetSdkVersion rootProject.ext.targetSdkVersion
         versionCode 1
