+++ dapps/v2Explorer/android/app/src/debug/java/com/walletconnect/web3modal/rnsample/ReactNativeFlipper.java
@@ -4,7 +4,7 @@
  * <p>This source code is licensed under the MIT license found in the LICENSE file in the root
  * directory of this source tree.
  */
-package com.v2explorer;
+package com.walletconnect.web3modal.rnsample;
 
 import android.content.Context;
 import com.facebook.flipper.android.AndroidFlipperClient;
