+++ dapps/v2Explorer/android/app/src/main/java/com/walletconnect/web3modal/rnsample/InstalledAppModule.java
@@ -1,4 +1,4 @@
-package com.v2explorer;
+package com.walletconnect.web3modal.rnsample;
 
 import android.content.pm.PackageManager;
 import android.os.Build;
