+++ dapps/v2Explorer/android/app/src/main/java/com/walletconnect/web3modal/rnsample/MainApplication.java
@@ -1,4 +1,4 @@
-package com.v2explorer;
+package com.walletconnect.web3modal.rnsample;
 
 import android.app.Application;
 import com.facebook.react.PackageList;
