+++ dapps/v2Explorer/android/app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">v2Explorer</string>
+    <string name="app_name">Web3Modal</string>
 </resources>
