+++ dapps/v2Explorer/app.json
@@ -1,4 +1,4 @@
 {
-  "name": "v2Explorer",
-  "displayName": "v2Explorer"
+  "name": "Web3Modal",
+  "displayName": "Web3Modal"
 }
