+++ dapps/v2Explorer/ios/Podfile
@@ -21,7 +21,7 @@ if linkage != nil
   use_frameworks! :linkage => linkage.to_sym
 end
 
-target 'v2Explorer' do
+target 'Web3Modal' do
   config = use_native_modules!
 
   # Flags change depending on the env values.
@@ -43,7 +43,7 @@ target 'v2Explorer' do
     :app_path => "#{Pod::Config.instance.installation_root}/.."
   )
 
-  target 'v2ExplorerTests' do
+  target 'Web3ModalTests' do
     inherit! :complete
     # Pods for testing
   end
