+++ dapps/v2Explorer/ios/Web3Modal.xcworkspace/contents.xcworkspacedata
@@ -2,7 +2,7 @@
 <Workspace
    version = "1.0">
    <FileRef
-      location = "group:v2Explorer.xcodeproj">
+      location = "group:Web3Modal.xcodeproj">
    </FileRef>
    <FileRef
       location = "group:Pods/Pods.xcodeproj">
