+++ dapps/v2Explorer/ios/Web3Modal.xcworkspace/xcshareddata/IDEWorkspaceChecks.plist
