+++ dapps/v2Explorer/ios/Web3Modal/Images.xcassets/AppIcon.appiconset/Contents.json
