+++ dapps/v2Explorer/ios/Web3Modal/Images.xcassets/AppIcon.appiconset/iOS App Icon_20pt@2x.png
