+++ dapps/v2Explorer/ios/Web3Modal/Images.xcassets/AppIcon.appiconset/iOS App Icon_29pt@3x.png
