+++ dapps/v2Explorer/ios/Web3Modal/Images.xcassets/AppIcon.appiconset/iOS App Icon_40pt@2x.png
