+++ dapps/v2Explorer/ios/Web3ModalTests/Info.plist
