+++ dapps/v2Explorer/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "v2explorer",
+  "name": "web3modal",
   "version": "0.0.1",
   "private": true,
   "scripts": {
